# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tech Dev API",
        "description": "API desenvolvida para o desafio da Gazin-tech",
        "contact": {
            "email": "caiogrechic50@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/niveis": {
            "get": {
                "tags": [
                    "Níveis"
                ],
                "summary": "Lista todos os níveis",
                "description": "Retorna uma lista paginada de níveis",
                "operationId": "ba69477b5febe5b79b7d473ca2300af3",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Campo que vai ser ordenado",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordenação crescente (asc) ou descrescente (desc)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Filtra pela descrição do nível",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de níveis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "nivel": {
                                                        "description": "Nome do nível",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida"
                    }
                }
            },
            "post": {
                "tags": [
                    "Níveis"
                ],
                "summary": "Cria um novo nível",
                "description": "Cria um novo nível",
                "operationId": "1b888cad19da322f4f972a17b927515e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "schema": "Nivel",
                                "title": "Nivel",
                                "properties": {
                                    "nivel": {
                                        "description": "Nome do nível",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nível criado"
                    },
                    "400": {
                        "description": "Requisição inválida"
                    }
                }
            }
        },
        "/api/niveis/{id}": {
            "put": {
                "tags": [
                    "Níveis"
                ],
                "summary": "Atualiza um nível existente",
                "description": "Atualiza os dados de um nível existente",
                "operationId": "8b8485890960c6f95f2331f1dbd379ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do nível",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "schema": "Nivel",
                                "title": "Nivel",
                                "properties": {
                                    "nivel": {
                                        "description": "Nome do nível",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nível atualizado"
                    },
                    "404": {
                        "description": "Nível não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Níveis"
                ],
                "summary": "Exclui um nível",
                "description": "Exclui um nível existente",
                "operationId": "fcdcfc181a7d1f9da3dcafd0d9602788",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do nível",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Nível excluído com sucesso"
                    },
                    "400": {
                        "description": "Nível associado a desenvolvedores, não pode ser excluído"
                    },
                    "404": {
                        "description": "Nível não encontrado"
                    }
                }
            }
        }
    },
...
}